feat(rust): ockam cluster attach
attach repl and other inlets to a running zone
